requirements.txt
alpha_vantage==3.0.0
Faker==33.1.0
fastapi==0.121.1
lxml==6.0.2
pandas==2.3.3
pymysql==1.1.2
python-dotenv==1.2.1
Requests==2.32.5
starlette==0.49.3
tqdm==4.66.2
yfinance
